+++ test/admin.test.js
@@ -1,94 +1,85 @@
-var run = require('child_process').spawn;
 var exec = require('child_process').exec;
 
-  describe('has admin functions', function(){
-    this.timeout(15000)
-    before(function(done){
-      //demo.admobilize@gmail.com // Password: admobdemo2016
-      exec('matrix set env dev')
-      done();
+describe('has admin functions', function(){
+  this.timeout(15000)
+
+  describe('can switch environments', function () {
+
+    it('`matrix set env local`', function(done){
+      fn.run('matrix set env local', {
+        checks: ['local'],
+        postCheck: function(done){
+          var env = fn.readConfig().environment.name;
+          if ( env !== 'local'){
+            done('Invalid environment')
+          } else {
+            done();
+          }
+        }
+      }, done)
     })
 
-    describe('can manage user state', function () {
-      // it('`matrix register`')
-
-
-      it('`matrix login`', function(done){
-        var loginProc = run('matrix', ['login']);
-        loginProc.stdout.on('data', function(out){
-          // console.log(out.toString())
-          if ( out.toString().indexOf('username') > -1 ){
-            loginProc.stdin.write('demo.admobilize@gmail.com\n')
-          } else if ( out.toString().indexOf('password') > -1){
-            loginProc.stdin.write('admobdemo2016\n')
-          } else if ( out.toString().indexOf('Login Successful') > -1){
-            // console.log(readConfig())
-            if ( readConfig().user.hasOwnProperty( 'token' )) {
-              done();
-              loginProc.kill();
-            }
-          }
-
-
-        })
-        // console.log(readConfig())
-      })
-
-      it('`matrix use`', function (done) {
-        var useProc = run('matrix', ['use 123456']);
-        useProc.stdout.on('data', function (out) {
-          var out = out.toString();
-
-          if ( out.indexOf('demo') > -1 ){
+    it('`matrix set env production`', function(done){
+      fn.run('matrix set env production', {
+        checks: ['production'],
+        postCheck: function(done){
+          var env = fn.readConfig().environment.name;
+          if ( env !== 'production'){
+            done('Invalid environment')
+          } else {
             done();
           }
-        })
-      })
-
-      // after('`matrix logout`', function (done) {
-      //   var loginProc = run('matrix', ['logout'])
-      //   loginProc.stdout.on('data', function(out){
-      //     // console.log(out.toString())
-      //     var f = false;
-      //     if (f === false && out.toString().indexOf('Logged Out Successfully')){
-      //       f = true;
-      //       done();
-      //       loginProc.kill();
-      //     }
-      //   })
-      // })
+        }
+      }, done)
     })
-    describe('can set an environment', function () {
-      it('`matrix set env local`', function(done){
-        var loginProc = run('matrix', ['set','env','local'])
-        loginProc.stdout.on('data', function(out){
-          if (out.toString().indexOf('Env: local') > -1 ){
+
+    it('`matrix set env dev`', function(done){
+      fn.run('matrix set env dev', {
+        checks: ['dev'],
+        postCheck: function(done){
+          var env = fn.readConfig().environment.name;
+          if ( env !== 'dev'){
+            done('Invalid environment')
+          } else {
             done();
-            loginProc.kill();
           }
-        })
-      })
-      it('`matrix set env production`', function(done){
-        var loginProc = run('matrix', ['set','env','production'])
-        loginProc.stdout.on('data', function(out){
-          if (out.toString().indexOf('Env: production') > -1){
-            done();
-            loginProc.kill();
-          }
-        })
-      })
-      it('`matrix set env dev`', function(done){
-        var loginProc = run('matrix', ['set','env','dev'])
-        loginProc.stdout.on('data', function(out){
-          if (out.toString().indexOf('Env: dev') > -1){
-            done();
-            loginProc.kill();
-          }
-        })
-      })
+        }
+      }, done)
     })
   })
 
-function readConfig(){
-  return JSON.parse( require('fs').readFileSync(require('os').homedir() + '/.matrix/store.json') );
-}
+  describe('can login', function (){
+    it('`matrix login`', fn.login );
+  });
+
+  describe('can make, list and delete devices', function(){
+    before('`matrix register device`', fn.registerDevice );
+    it('`matrix list devices`', function(done){
+      fn.run(
+      'list devices',
+      {
+        checks: [
+          'test-device'
+        ]
+      }, done);
+    })
+
+    // doesnt work right now
+    it.skip('`matrix remove`', function(done){
+      fn.run('remove test-device', {
+        responses: [
+          ['test-device', 'y\n']
+        ],
+        checks: [
+          'Device successfully removed'
+        ]
+      }, done)
+    })
+
+
+  });
+
+  describe('can logout', function(){
+    it('`matrix logout`', fn.logout);
+  })
+});